delete canReadGameFiles method
What this function does is: Take loaded scripts and try to reload the last one again. If it cannot succeed, it shouts an error. However, it's completely redundant, because if it can't load that file, it can't load rpg_core.js either, so this function will never execute when it can't execute and whenever it executes, it confirms what we already know from loading other scripts.